definitions/raster-calculations/containers/process-dem/requirements.txt
numpy
python-decouple